fix[codegen]: cache result of iter eval (#4488)
prior to this commit, multiple evaluation of a single expression is
possible in the iterator target of a for loop. while the iterator
expression cannot produce multiple writes, it can _consume_ side
effects produced in the loop body (e.g. read a storage variable
updated in the loop body) and thus lead to unexpected program
behavior. specifically, reads in iterators which contain an `IfExp`
(e.g. `for s: uint256 in ([read(), read()] if True else []))` would
issue one evaluation of the list `[read(), read()]` per loop iteration,
thus interleaving reads with writes in the loop body.

this commit fixes the issue by using `cache_when_complex` to enforce
evaluation of the iterator before entering the loop body.

this is incidentally also a performance fix, since it moves the
repeated evaluation into a one-time evaluation.

references:
- https://github.com/vyperlang/vyper/security/advisories/GHSA-h33q-mhmp-8p67